Actualizar services.dart, entities.dart y 7 más archivos...

+++ lib/user_data/model/entities.dart
@@ -124,7 +124,7 @@ class Feeding extends UserDataEntity {
 
 class Activity extends UserDataEntity {
   final UserDataValueObject activityType;
-  final double minutes;
+  final int minutes;
 
   Activity({int id, DateTime eventDate, int userId, String entityType, this.activityType, this.minutes}) : super(id, eventDate, userId, entityType);
 
@@ -147,7 +147,7 @@ class Activity extends UserDataEntity {
 
 class InsulinInjection extends UserDataEntity {
   final UserDataValueObject insulinType;
-  final double units;
+  final int units;
 
   InsulinInjection({int id, DateTime eventDate, int userId, String entityType, this.insulinType, this.units}) : super(id, eventDate, userId, entityType);
